part2/pyproject.toml
[project]
name = "part2"
version = "0.1.0"
description = ""
authors = [{ name = "danielAdama", email = "adamadaniel321@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "langgraph (>=0.6.6,<0.7.0)",
    "langchain-core (>=0.3.75,<0.4.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tenacity (>=9.1.2,<10.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "prometheus-client (>=0.22.1,<0.23.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
